finish the assgniment

+++ app/src/main/java/weather/CloudsEntity.java
@@ -3,5 +3,5 @@ package weather;
 public class CloudsEntity {
     @com.google.gson.annotations.Expose
     @com.google.gson.annotations.SerializedName("all")
-    public int all;
+    public float all;
 }
